backend/requirements.txt
Django>=4.2.7,<4.3.0
django-cors-headers>=4.3.0,<4.4.0
djangorestframework>=3.14.0,<3.15.0
django-extensions>=3.2.3,<3.3.0
pytest>=7.4.3,<7.5.0
pytest-django>=4.7.0,<4.8.0
pytest-cov>=4.1.0,<4.2.0
requests>=2.31.0,<2.32.0